requirements.txt
scikit-learn==1.6.1
xgboost
pandas
numpy
category-encoders
fastapi
streamlit
uvicorn
kafka-python
watchdog
